+++ cubemx/USB_DEVICE/App/usb_device.c
@@ -65,6 +65,19 @@ void MX_USB_DEVICE_Init(void)
 {
   /* USER CODE BEGIN USB_DEVICE_Init_PreTreatment */
 
+  // we do not have VBUS sensing on PA9, the sensing is handled by PE1 in software
+  __HAL_RCC_GPIOE_CLK_ENABLE();
+
+  GPIO_InitTypeDef GPIO_InitStruct = {};
+  GPIO_InitStruct.Pin = GPIO_PIN_1;
+  GPIO_InitStruct.Mode = GPIO_MODE_OUTPUT_PP;
+  GPIO_InitStruct.Pull = GPIO_NOPULL;
+  GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_LOW;
+  HAL_GPIO_Init(GPIOE, &GPIO_InitStruct);
+
+  // disable USB pull-up power source
+  GPIOE->BRR = GPIO_PIN_1;
+
   /* USER CODE END USB_DEVICE_Init_PreTreatment */
 
   /* Init Device Library, add supported class and start the library. */
@@ -86,6 +99,9 @@ void MX_USB_DEVICE_Init(void)
   }
 
   /* USER CODE BEGIN USB_DEVICE_Init_PostTreatment */
+
+  // enable USB D+ pull-up power source, the pull-up becomes active only if a USB cable is connected to the device
+  GPIOE->BSRR = GPIO_PIN_1;
 
   /* USER CODE END USB_DEVICE_Init_PostTreatment */
 }